Get sheet music justification to work
Doesn't work if repeat ends at the end of a staff (or if there is more than one repeat barline)

Doesn't work if time signature changes

+++ src/main/java/GUI/DrawNotes.java
@@ -22,9 +22,9 @@ public class DrawNotes {
 	private Pane pane;
 	private double x, y, yStaff;
 	private Note note;
-	private int noteSpacing;
+	private double noteSpacing;
 
-	public DrawNotes(Pane pane, double x, double y, Note note, String instrument, double yStaff, int noteSpacing) {
+	public DrawNotes(Pane pane, double x, double y, Note note, String instrument, double yStaff, double noteSpacing) {
 		this.pane = pane;
 		this.x = x;
 		this.y = y;
